backend/mdv/observability.py
from __future__ import annotations

import os
from typing import Optional

import sentry_sdk
from opentelemetry import trace
from opentelemetry.exporter.otlp.proto.http.trace_exporter import OTLPSpanExporter
from opentelemetry.sdk.trace import TracerProvider
from opentelemetry.sdk.trace.export import BatchSpanProcessor

from .config import settings


def init_observability(service_name: str) -> None:
    # Sentry
    if settings.sentry_dsn:
        sentry_sdk.init(dsn=settings.sentry_dsn, traces_sample_rate=0.1)

    # OpenTelemetry
    if settings.otel_exporter_otlp_endpoint:
        provider = TracerProvider()
        processor = BatchSpanProcessor(OTLPSpanExporter(endpoint=settings.otel_exporter_otlp_endpoint))
        provider.add_span_processor(processor)
        trace.set_tracer_provider(provider)

